requirements.txt
numpy==1.24.3
h5py==3.8.0
tqdm==4.65.0
Pillow==9.5.0
opencv-python==4.7.0.72
pandas==1.5.3
einops
timm
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121